started to update ip server to use flask
