+++ intent_parser/requirements.txt
@@ -1,3 +1,4 @@
+Flask==1.1.2
 google-api-python-client
 google-auth-httplib2
 google-auth-oauthlib
